pyproject.toml
[project]
name = "jcode-plans-py"
version = "0.1.0"
description = "Plan file storage extracted from jcode"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jcode-conf-py>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jcode_plans"]

[tool.pytest.ini_options]
testpaths = ["src/jcode_plans/tests"]

[tool.uv.sources]
jcode-conf-py = { git = "https://github.com/junix/jcode-conf-py.git" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
